Test and modify authentication logic and much more
The config field `credentials` has been renamed `authentication`. Now
authentication can be disabled in the config file with the
`authentication.disabled` field

Simple authentication no longer can be created with an empty username
and password. By extension, when `authentication.disabled` is false,
there must be a non-empty username and password in the `authentication`
config section.

Added errors.Wraps as a convenience method to wrap a string as an error.

The handler environment now has a field for a clock. This clock is used
by handlers to get the current time, and can be used in tests to
provide a constant timestamp.