Fix bug where articles with the same name but different file extensions would override eachother
This was made apparent when writing and saving articles using an editor
like Vim. Some editors have a behave where they first create a temporary
file with the same name as the file you're editing, except it has some
extra characters following it. For example if you're editing `about.md`,
it will make a temporary file named `about.md~`. These files were
colliding in the article Map since the route to the article was being
used as a key, which doesn't include the file extension.

To fix this problem, articles are now stored in the map using the
relative path to the article as the key, which inclues the file
extension. To ensure that routes still don't require any file extension
to access an article, only the ".md" extension is supported for
articles.